Move builder tests
